Modular pulse shape compared to the RX gate the IBM QuantumPlatform
The code & data related to the Appendix A. 

'gaussian_pulse_gate_ibmq_mumbai_q0' contains the code related to using a modular pulse to charge a quantum battery on an IBM Quantum Platform.  
The work of the paper "IBM Quantum Platforms: A Quantum Battery Perspective", Gemme et al., https://doi.org/10.3390/batteries8050043
was used as the inspiration for this code. 
The tutorial on Qiskit Pulse by Qiskit was used to write the code. 

'rx_gate_ibmq_mumbai_q0' contains the code to charge a quantum battery on the IBM Quantum Platform using the built-in RX gate.